Add function stubs for FERC1 DB table ingestion.
Created stub functions for the ingestion of each of the initial FERC1
tables, and generally added control logic to allow various parts of the
database definition, creation, and population can be controlled separately,
which should make for easier testing. Also added the option for some
verbose output so we can see what's going on.