Debugged f1_hydro import and standardized filenames
We noticed that the plant_name fields from FERC1 were sometimes showing
up as different just because of capitalization or leading/trailing
white space. This causes errors, because they are part of the foreign
key defining FERC plants (along with respondent_id).  To get around this,
every plant name is now forced to be Title Case and leading/trailing
white space is stripped whenever those fields are brought into the PUDL
database.

Debugged the import of f1_hydro, which had gotten messed up by a poor
merge. Changed some of the DB field names to reflect units, converted
units to MW and MWh on ingest to match our other tables.  Switched
data type for monetary fields in f1_hydro to be Numeric(14,2).

Also fixed an error where the name of an EIA ORM class wasn't matching up
with the class as it was being called (...EIA vs. ...EIA923)