Fixed bugs preventing repeated DB initializations.
On the 2nd or subsequent initialization of the FERC Form 1 DB, the
initialization would fail, claiming that f1_respondent was being
re-defined as not a primary key.  This seems to have been due to a stale
MetaData object persisting from the previous initialization. Before we
construct the new MetaData object it now runs ferc1_meta.clear() to remove
all the existing table definitions.

Once that bug was fixed, I found that the ferc_electric_plant_accounts
DataFrame defined in constants.py no longer had descriptions in it on
2nd and later database initializations. This is because ferc_accts_df
was being set to the constants DataFrame itself, rather than a copy of
the data frame in the ingest_static_tables() function... and then the
descriptions were getting dropped in place, before merging with the
Plant In Service table (to replace line numbers with FERC Account IDs.).

Also wrapped some overly long lines.