Set up basic test cases for pudl, ferc1, & eia923.
Now that we've got a bit of a code base, it's time to start exercising it
in an automated way.  I have set up three sets of tests, one each for
the pudl, ferc1 and eia923 modules. The ferc1 and eia923 tests try and
import all data which we expect to work into their respective intermediary
forms. This means bringing years 2004-2015 of FERC into the ferc1 postgres
database, and years 2014-2016 of all pages of the eia923 data into pandas
data frames.  The pudl test just tries to bring a single year (2015) into
PUDL, to exercise all of its functionality without taking a huge amount of
time.

Overall the test suite takes about 11 minutes to run.  You will need to
have a ferc1_test and pudl_test database created locally, and the name of
the working pudl database has been changed to pudl rather than
pudl_sandbox.  To run the tests, from the top level pudl repository
directory, run:

 pytest -s

PyTest is a standard part of the Anaconda distribution and should already
be installed.

Please run the test suite after debugging your code locally, and before
pushing new code up to the shared remote repository.

We should also come up with tests that check the actual integrity of the
database and data -- things we can do with the data once its in the
database that have known outcomes we can check.  If you think of any such
more detailed tests, let's talk about them!