Added glue for integrating new EIA923 ingest funcs
When a new ingest function is added, in order for it to be tested
automatically, it needs to be added to the list of (source)_pudl_tables in
the constants module -- otherwise it will not be ingested by default. I
added a list of the remaining EIA923 ingest functions there, but commented
them all out... since they don't work yet.

Additionally, the new ingest function needs to actually be called within
init_db() in pudl.py.  I added all of the stub functions to th dictionary
of tables:ingest_funcs there so that they'll automatically be run *if*
someone adds their new table to the list of supposedly working tables in
constants.

After running tests for ingest_boiler_fuel_eia923, and seeing them fail, I
removed it from the eia923_pudl_tables list, pending more debugging by
Alana.