Add the column mapping for generator tab in eia923
In order to stitch together the historic columns from eia, this is
a new dataframe to index into which maps the cleaned column names
to the column names that exist in this tab for each year. The
dataframe was added into constants to be used to populate a
dataframe for all the years.