Clean up docstrings and doctstring formatting for eia923.py, ferc1.py, pudl.py
I'm not sure how deeply we need to explain relatively straightforward functions so I tried to focus on the ones with more arguments. I mainly stuck to explicating what was already annotated and making sure that it was formatted correctly as well as key arguments. If we need to do more work explaining what's returned and what all the arguments are in every case I'm happy to help.