Integrate CSV loading for EIA923 data
The full set of data from eia923 consists of millions of records...
which it turns out takes a while to load. Just 2011-2015 eia923 plus
2007-2015 data from FERC was taking more than an hour. Much of that
time was spent in DataFrame.to_sql() which is (it turns out) notoriously
slow for loading data into DBs. Native CSV file loading is much faster,
and dumping a pandas DataFrame.to_csv() is also quite fast, so I changed
the code to dump the eia923 dataframes out to CSV files, and then
directly load those CSV files into the db using a little python module
called postgres_copy (which the project now depends on).  This also gives
us an intermediary and ready-to-import version of the data that goes into
the database, which will probably be useful functionality in the long run.

In the course of making these changes, I also reconciled the column names
that are used for re-naming old eia923 spreadsheet columns to match the
database columns, rather than the 2014-2016 EIA923 spreadsheet column
names (why have 2 renamings, when you can just have one?).

I also changed get_eia923_files to take a single year and return a single
file, rather than taking a list of years, and return a list of files...
since that's more the usage mode for the ExcelFile dictionary generation.

Resolves: #79, #66,