Notebook looking into FERC average heat content and fuel cost per mmbtu data for coal and gas.
This notebook attempts to identify outlying populations for ratios like cost per mmbtu and average heat content, where we feel like we have a good idea of what the 'normal' range of values should be. For fuel cost per mmbtu and average heat content there are functions to try to bring in as many outlying populations as have been found so far.