Add scripts directory and a PUDL DB init script.
In order to make it easy for folks who just want to use the PUDL DB w/o needing
to really mess with the code or notebooks necessarily, and to have the DB get
initialized from a clean environment, I've created an initialization script.

The scripts/ directory is meant to hold other stand alone scripts which utilize
the PUDL modules, but are not part of the package itself, and which are not
part of the data analysis or processing (which would probably go under
results/.

Because the pudl package isn't yet getting "installed" to the site-packages
directory, there's still a hard coded addition to the python import path
inside the script -- we will need to fix that with a relative import and/or
get the modules installing in the site-packages directory eventually.