Fix pumped_storage table definition & ingest.
Due to a typo in the __tablename__ in models_pudl.py for the pumped
hydro table, none of the processing taking place in the ingest function
was actually being applied to the table which had been defined using
SQL Alchemy.  After fixing that typo, the ingest_pumped_storage_ferc1()
function needed to be debugged and a few data issues worked through.

Also added more extensive docstrings for several of the ingest functions.