Fix bugs related to ingest of EIA923 for 2009-2010.
EIA923 data from 2009-2010 and 2011 and later was being stored in the
spreadsheets as different types of data, which was causing drop_duplicates()
not to remove some records which were in fact duplicates.  This was especially
a problem in the boilers_eia923 and generators_eia923 where boiler_id and
generator_id were usually integers, but strings in some cases. The solution was
to cast them all to the string type before calling drop_duplicates(). There may
be other instances in which this type heterogeneity is resulting in duplicate
entries being pulled into the PUDL DB, but we aren't noticing because they
aren't in fields which prohibit duplicate records -- this was the case with the
boilers_eia923 table -- it had a surrogate key instead of a composite key, and
there were duplicates in the (plant_id, boiler_id) tuple, but because that
wasn't the primary key, there was no problem on ingest, whereas in the
generators_eia923 table, we *were* using (plant_id, generator_id) as a
composite primary key, which is why that table was failing to ingest when years
from both before and after the 2010/2011 boundary were being pulled.

In the course of debugging I decomposed all the ingest functions such that they
each only populate a single table in the database, and made the process for
ingesting the plant_info_eia923 table more similar to all of the other tables
that come from the EIA923 spreadsheets.