OMG 2016 WORKS!
Okay. The main alteration that happened here was modifying
get_eia923_plant_info to replace empty fields from the plant
frame tab (i.e. a plant id that did not show up in the plant
frame tab for whatever reason) with info from the other tabs.
The ingest_plant_info_eia923 was altered to account for adding
the nameplate capacity.

The second major change came through adding the missing 2016
eia923 plant ids into the mapping spreadsheet (notebooks that
generated this list to be checked in shortly).