Simplify imports in pudl, consolidate cleanstrings
We have been accumulating a huge list of individually imported names
at the beginning of the pudl module. This is unwieldy and also makes
it hard down in the code to see where a particular function or variable
is ultimately coming from. All of the individual imports have now been
replaced with module or package level imports, some using shorthands
(like import sqlalchemy as sa) to keep variable names from getting too
long. Similar cleanups of the imports in other modules will also need
to happen.

Additionally, the cleanstrings() function which used to live in the
ferc1 module was brought into the pudl module, because a very similar
kind of thing needed to be done to some EIA data as well.  A duplicate
of the cleanstrings function was added to the eia923 module. That function
was removed, and they ought to all refer back to the pudl.cleanstrings()
function now. All canonical names are getting lower-cased... which meant
I needed to change the list of aer_fuel_type values in constants.py to
be consistent.  We might want to have a preserve_case=True|False flag in
the future, for instances where the case actually matters in the codes.