Added capacity into plant_info_eia923 table
Generated a new df in the get_eia923_plant_info formula which just
pulls the plant id, capacity and year, and filters it for only the
records which contain values for the capacity (the 2011 records).
This df was merged into the plant_info_complied df like the rest
of the df for each tab. The field in the data table definition was
to include the capacity, as well as the subset of columns used in
the ingest function.