Fix multi-year yearly to monthly data conversion.
Found and fixed a bug that was causing the conversion of annual EIA923
records (with 12 monthly values all in the same record) to monthly
records (with a single month in each record) to merge incorrectly when
multiple years worth of data were present, creating duplicate, wrong
records. This also inflated the number of records by a factor of N
where N is the number of years being loaded.