Use official EIA923 fuel heat content bounds.
In the cleaning routine that forces fuel heat contents to be within
specific physically reasonable bounds (and multiplies some data to
bring it within those bounds if it's outside because of unit errors
in the data entry) we switched to using the bounds that are publicized
by EIA themselves.  I.e. 9-30 mmBTU/ton for coal and 0.8-1.2 mmBTU/mcf
for natural gas.