Connect EIA923 coal mine & fuel receipts tables
Because the coalmine_info_eia923 table uses surrogate keys, and because
of the way we are populating the tables (using postgres_copy rather than
the ORM) there's no easy way (that we've found) to automatically add
the coalmine_id into fuel_receipts_costs_eia923 when it is populated so
that the two tables can be linked to each other with a foreign key and/or
a relationship.

The solution implemented here uses all of the relevant data fields in the
coalmine_info_eia923 table as a "natural key" to merge a dataframe of
coalmine info with the fuel receipts & costs info (bringing in the
surrogate keys from coalmine_info_eia923) while the fuel receipts and
costs data is still a fresh dataframe that's just been read out of the
EIA923 spreadsheets (and thus still has the coalmine info in it). Then
the coalmine info data fields are dropped, before the FRC table is
populated.

The coalmine info is pretty aggressively deduplicated.  All the coal mine
names are put in Title Case, whitespace is stripped and compacted, and all
non-alphanumeric characters are removed.  Additionally, only one entry is
retained with each MSHA ID, even if some of the other fields (like the
coalmine name) are different between the records... since if the MSHA IDs
are identical, they really should be referring to the same mines.

The coalmine_county field was also changed into an Integer rather than a
String.  The only non-Integer entries in it were "IMP" or "IM" denoting
coal which had been imported. Those values are now NULL, since there is
no true "county" associated with them, and the coalmine_state field has
a country code in it, which implies the coal was imported.  We should
really split coalmine_state into two fields... coalmine_country (mostly
USA) and coalmine_state... but, that's another problem.

Also commented out the plant_ownership ingest function call, because it
was breaking the data import.

Resolves: #89