EIA860 table added (boiler/generator associations)
In order to pull in additional information from EIA860, additional
modules have been added. A class definition for a
boiler_generator_assn table was defined, then functions to create
the data directory, file path, xlsx objects, column map dataframe
from a selected tab. In pudl.py, an ingest function was defined
for this table and the initialization function was updated to
include eia860.