Create function for finding new FERC plants
FERC released their 2016 Form 1 data, and we need to integrate the
new plants that appear within it into our EIA to FERC mapping. I
added a new function analysis.fercplants() that enables grabbing
this kind of information -- all the new plants and/or all the FERC
plants from the several tables that contain that kind of information.

Also revised the old notebook which we used to do this kind of thing
before, to use the new function (and output an XLSX file).  May need
to do a similar thing to get all the new utilities, though technically
that information is contained within this data as well.