Create more extensive FERC plant list for diffing
Extended the years which are used to enumerate all FERC plants back to
2004, which is the earliest version of the DB which will ingest today.
Changed notebook to re-initialize the FERC DB before pulling the list.

Removed the old ferc1_all_plants_2004-2015_id_mapping.ipynb because it
has been superceded by the current funciton/notebook.

Told git to ignore the .xlsx files which are output.