Extend FERC data back to 2004, re-org pudl.init_db()
The only thing that was keeping us from having FERC data that goes
all the way back to 2004 was the fact that the line_no mapping for
the f1_plant_in_srvce table breaks down before 2007 (when a new line
got inserted, offsetting all the line-to-value correspondence). Because
seeing the historical fuel & O&M costs for FERC is now a priority, I
just told the plant_in_service ingest function to ignore any data from
before 2007, and all the other tables we're pulling in work just fine.
I updated the constants and some default values to reflect this.

I also cleaned up the pudl.init_db() function a little - creating a
little wrapper function for each of eia923, eia860, and ferc1 so that
the init_db function is easier to read.  And so we can more easily
re-order the ingest of those three sets of data on the fly, which can
be nice for testing (i.e. you can put the thing that is breaking first).
This should also allow simpler individual testing of the components in
Jupyter notebooks.