Added 2016 FERC Form 1, EIA923 data to pull
Community Energy wanted the 2016 data, which is now out for both
FERC Form 1 and EIA 923, but not for EIA860. The FERC Form 1 pull
was easy, but the EIA sheet is a combination of EIA860 and EIA923
data.  To get the EIA923 data reported, I replicated the 2015 EIA860
fields (which don't really change year to year -- plant capacity,
technology, location, etc.) and merged them with the 2016 EIA923
data, to give some semblance of a full dataset for 2016.