Document data zipper and add test case
Gave much fuller comments and docstrings to the datazipper functions,
and created a test case that should run quickly, to make sure we don't
break the zipper with changes to the code elsewhere.  Still needs to be
made much faster, and generalized to work with other kinds of data that
need to be zipped.