Migrate notebook functions into analysis module
Took several functions that were used in the notebook for exploring
FERC to EIA plant correlations for several variables, and migrated
them in a slightly more general form into the analysis module.  Now
they work with not just the simple PUDL plants (one generator/plant on
both the EIA & FERC sides) but also more complicated plants, aggregating
whatever variables you're interested in by the entire plant.

The tables that are being aggregated include:
 - fuel_ferc1 (fuel heat content)
 - steam_ferc1 (fuel costs and net generation)
 - fuel_receipts_costs_eia923 (fuel costs by fuel)
 - generation_fuel_eia923 (net generation & total mmbtu by plant)

There are also some functions which do queries and joins from the database,
giving richer information to work with (mainly more ID values for
merging with other datasets, and some more names).

In order to make this stuff work on a per generator basis, we need the
other MCOE data to correlate, on a per generator basis.