Moved table compilation routines to output module
The analysis module has been a collecting area for useful functions we
want to have access to elsewhere, but it's become a bit too large. Starting
the process of re-organizing some of that content into thematically
similar modules instead. The output module is going to be a library of
re-usable pulls from the database, either for further manipulation and
analysis or for handing off to others in tabular form.

This commit moves functions for pulling:
 * fuel_receipts_costs_eia923
 * generators_eia860
 * fuel_ferc1
 * plants_steam_ferc1
 * generation_fuel_eia923

 And also updates the generators_eia860 and fuel_receipts_costs_eia923
 pulls to be, uh, correct -- uses several simple select statements to get
 entire tables, which are then joined together using pandas dataframes.