Standardize date formatting for EIA860 generators
Integrated the use of Date fields for all year/month pairs in the
EIA860 Generators table, using a function in clean_pudl that finds
matching pairs of fields of the form BASENAME_year and BASENAME_month
and converting them to a real Date object. This required changing a
couple of columns in the constants.py file to use _month and _year
suffixes rather than prefixes.  It also required changing the DB model
for the EIA860 generators table -- and many of these changes should
probably be propagated elsewhere too -- getting rid of all the dangling
year/month pairs, using real dates, and naming the columns appropriately,
with _date for a Date object, and _year for bare years (which we use
all over the place for report_year)