Comments from code review of MCOE module & calcs.
Moved heat_rate() and gens_with_bga() from analysis.py to mcoe.py.

Reviewed all of the functions that go into the calculation and attribution
of fuel consumption and costs on a per-generator basis.  Wrote docstrings
and added comments (in the docstrings) about stuff that might be good to
check on or clarify.

Only made very minor code modifications, on syntactically questionable
stuff like using == True instead of a direct mask in dataframe selections.

Updated the test_mcoe() function to reflect the new location of the
heat_rate() function, in mcoe.py