Update heat_rate to deal with boiler/gen associations.
Debugged some issues with the heat_rate calculation that had to do with
incorrectly identifying some plants as "complete" that really weren't,
and missing boiler_fuel_eia923 data -- fuel consumption that's simply
not present in the data.  Still lots of issues to work out (and need to
convert them to actual issues in GitHub) but we have something that seems
to at least kinda work now... Made some nice histograms of generator heat
rates in the notebook. Updated tests to work with the new functions.