Allow specification of years for MCOE data pulls
Unfortunately, all of the EIA data that we have isn't ready for
use in calculating the heat rates and thus fuel cost components.
Only data from 2014 and later appears to be usable as-is.
It's also interesting to be able to look at the output on a year
by year basis, to see how fuel costs have changed over the years.

These changes allow a list of years to be specified in the data pulls
that go into creating the heat_rate and fuel_cost estimates on a
generator by generator basis.

Also a little bit of cleanup related to which fields remain in the
heat_rate data frame after the calculation is done.