Do some string-mapping for EIA860 energy sources.
Added a call to cleanstrings() to map the complicated energy sources
in the generators_eia860.energy_source_1 field to our simplified field.
Really this should be applied to all of those energy_source_* fields.

Also updated the clean_pudl.cleanstrings() function to give the option
to "simplify" (i.e. strip and compress whitespace, and lowercase
everything) in both the dictionary of codes and the values coming out
of the raw data.