Simplify & standardize time-based grouping & merging
Created a function which allows the merging of two dataframes, one of
which has a date field, and another of which has an integer year field,
based on the year in the date.

Updated calls to deprecated pd.TimeGrouper() in favor of pd.Grouper().

Simplified some date/time manipulations within functions.