Allow flexible time-based aggregation of data pulls
For all the EIA 923 data pulling functions, allow the data to be pulled
either raw (with all records included) or aggregated by time.  If
aggregated by time, appropriately sum bulk fields, and re-calculate
per-unit fields on a weighted average basis -- but in this case we lose
the non-ID (groupby) and non-numerical (coded) fields.  Some of those
fields could in theory be added back in, if the code values can be looked
up as derived values which depend only on the ID fields (plant, boiler,
generator, etc.)  But as of now, that's not being done.