Standardize report_date, report_year, report_month
Changed tab maps and internal function calls to use report_ prefix
for all time-based columns which refer to when the data in that
structure was collected, rather than having some bare 'month' and 'year'
columns floating around, which can be confusing, given that there are
sometimes several different times being described in a given table.

Changed the convert_to_date() function to make a copy of the dataframe
which is passed in, modified, and eventually returned, rather than
quietly doing in place operations.

Also created a pudl_tables dictionary in constants.py, keyed by
data source, indicating the names of the tables which ought to be
populated by default when the database is created.