Update EIA860 to use report_date not report_year
Changed all the database table definitions for the EIA860 data
to use a report_date column of type Date, and to populate that
column using beginning-of-year dates. This makes them consistent
with the EIA923 data tables. Previously only one of the EIA860
tables was using date, and the rest were still year.