Integrate new fuel_cost function
Removed the last of the MCOE data pulls, and re-wrote the fuel_cost
function to assume that all generators & boilers are fully associated,
in anticipation of the new boiler generator mapping function that
Christina is creating. That function will eventually replace the
gens_w_bga and boiler_generator_pull_eia860 functions that remain here,
and be integrated into the database initialization process.