Add more robust test for annual data frequencies.
In some cases we need to verify that the dataframe
being analyzed has data which is compatible with annual
frequency -- when merging annual with sub-annual data in
merge_on_date_year() and when using extend_annual() to create
earlier generator & plant records for EIA860.  These changes
allow those frequencies to be inferred even when only a single
or two annual dates are present in the dataset.