Fixing some of MCOE tests.
Reintegrating the new inputs into the MCOE tests. Incorporated a start
and end date into the tests for all modules which reference the date
range in contants, integrated that into the output tests. The MCOE tests
aren't fully functional: the test_fuel_cost and test_mcoe_calcs. The
fuel cost function is failing bc of the fuel_type_pudl but Zane is (I
believe) working on this. The tests for capacity factor and heat rate
work and the fuel cost should work once its fixed.  We need to
incorporate tests on the content.