Move merge of fuel_type_pudl into heat_rate()
In order to calculate fuel costs, the heat_rate dataframe needs
to contain information about a generator's primary fuel -- what
kind of fuel it uses, and whether it's part of a single or
multi-fuel generator unit, for fuel cost attribution purposes.

Previously this information was being merged into the data frame
resulting from the heat_rate calculation in the top level mcoe
calculation function. For the purposes of testing and having the
heat rate input be ready to go into the fuel_cost calculation
immediately, that merge has been moved into the heat_rate calculation
instead.