Fix merge bug in heat_rate, update test suite.
The heat rate and fuel cost calculations were generating dataframes
with many more than one record per generator per reporting period, which
seemed really wrong -- in some cases there were as many as 24 records per
generator. It turned out this was due to gens_assn in mcoe.heat_rate()
not having unique per-generator records. Initially it includes all the
boiler generator associations, and a bunch of tags indicating how the
associations were determined, and the status of the association, but we
only need it to tag the generators as to whether they were ever part of a
completely associated plant, so the boiler ID was getting dropped, and the
records were being merged in with the other generation information.  But
there was still one record for each of the boiler_id values in the
dataframe. Whoops! Dropping duplicates on gens_assn after dropping the
boiler_id appears to fix this issue, and gives very clean outpus for
capacity factor, heat rate, and fuel costs on a per-generator basis.