Added generation unit ids to the BGA dataframe
For many applications the most useful sub-plant granularity
is the generation unit. This is a combination of boilers and
generators whose heat input and electricity output cannot be
separated, because they share infrastructure. The additions to
the mcoe.boiler_generator_association() function use network
analysis to infer which groups of boilers and generators are
connected to each other, and assigns them each a unique unit_id
within each plant. The new column is called unit_id_pudl and it
still needs to be integrated into the rest of the MCOE
calculations.

Note: there were several odd EIA plants that appear to have the
same boilers or generators associated with different unit_code
values throughout the year.  These are flagged when the function
is run, so we don't forget to fix them...