Re-organize output functions & MCOE calculation.
This ridiculously large commit (sorry) revises the MCOE calculations to
use generation units, rather than individual generators, for calculating
the heat rates. This simplifies the overall calculation substantially,
and will be useful when it's integrated into the core of the PUDL DB.

It also switches to using and object oriented PUDL Output class...  which means
rather than passing around a selection of several data frames for the MCOE
calculation, and trying to keep track of which is what, and the setup work
required to calculate the various derived values, we can use an output class
(outputs.PudlOutput) that knows how to generate those dataframes, as needed,
based on a few inputs set at the initialization of the object (frequency, start
& end dates, whether it's using the testing database).  The object thus
instantiated then pulls the dataframes it needs using the output functions, as
they are requested, and they persist as data members within the object.

This also makes testing much simpler, since we can create a fixture that
generates output objects having a given frequency and start/end dates, that
gets re-used over and over again.

This is a pretty major change from how we've been doing things, but it cleans
the code up dramatically, and should be extensible going forward, as we add
additional outputs and data we want to pull as we go.  It might be able to
accommodate graphical outputs too.