Integrate cleaning of ownership fractions in EIA860
Prior to 2012, the EIA860 generator ownership data is reported as
percentages (0 to 100), rather than fractions (0 to 1). This cleaning
function converts the earlier data to fractions before ingesting it
into the database.

Also integrates a check for completeness into test_eia860_output(),
which allows no more than 1 in 1000 generator-years to have less than
98% of its ownership described. There are a few dozen records that
have no ownership data (NaN), and several others that are 0.95 or 0.5
or 0.25 -- why their ownership descriptions are not complete is unclear,
but given that there are 20,000+ records, this doestn' seem like a big
issue.