Migrate ownership_eia860 transformation
Working on Issue #132. Trying out one table at a time. All of the
tranformation of the ownership table was migrated out of the ingest
function in init.py and now lives in transform.eia860. Right now, the
clean formula still outputs a dataframe which is used for the
csv_dump_load in the ingest function. Need to create a stremlined load
function to input all of the cleaned dataframes (or a transformed
dictionary).