Integrating extract and load process
Working on Issue #132. Using the ownership_eia860 table as an example to
fully extract the individual table ingest functions as a part of the
ingest process. The ingest_ownership_eia860 function is still in the
init module, but no longer a part of the process.

Now, ingest_eia860 has three main steps: make the dictionary of
dataframes from the spreadsheet, run the transform functions to make a
dictionary of transformed dataframes, then load them into tables.

I've moved a copy of _cvs_dump_load into a new load module and created a
small function for 860 that uses _csv_dump_load to load the tables from
the transformed datafames.