Migrating 923 into new transform/load process
Working on Issue #132. I removed all of the 923 ingest functions. All
the transofrmation now happens inside of the transform module and the
loading happens after creating a transformed dataframe of all of the
tables.

One quirk: the FRC table requires the auto incremented id's from the
coalmine table so we need to load the coalmine table which requires the
pudl_engine, so this function has differnt inputs than the other
transform functions.