Comment out buggy backfilling code.
The EIA 860 backfilling code appears to be wiping out a bunch of
report date information, resulting in (for example) all of the
generators_eia860 records being associated with only 2011 and 2012.
Commenting it out appears to avoid these issues for now. Checking it
back in as is, in case @alana wants to take a look and try to resolve.