ferc small plant clean name integration
There are plant_name_clean, which was previously only names that had
been standardized by hand in a csv. In working on Issue #132, I was
trying to find all the records of ferc plant names and this came up as
an issue. I first tired to remove the dirty plant names but ran into a
database relationship issue (bc there are no plant ids, we use the
respondent and plant name as the composit id, but the name used is the
dirty names which are currently generated from the id_mapping process.

So, I made the plant_name_clean column a full list of names but did not
remove the 'plant_name' which is the raw uncleaned names. Working on
Issue #152.