Create plant and gen entity tables
Now we have plant entity tables and generator tables. These are created
from the transformed dataframes to create these entity dataframes. They
are ingested first bc these are the primary location for the plant and
generator ids (or will be for the case of the generator ids).

I've also created annual info tables which will be used to normalize
plant/gen annual data. These are treated as data tables on injest, which
cased me to create a unified eia_transformed_dfs.