Generating FERC plant time series, CPI vs. sklearn
A couple of notebooks related to extracting usable FERC plant data
from the FERC database. Uses sklearn feature vectorization and similarity
metrics. Several parameters need to be optimized to best retrieve the
data, and these parts need to be integrated into that larger framework
to make the system robust. Still a work in progress.