Move fix_int_na and fix coalmine id
Fix_int_na was previously all being done in the transform step. This was
mucking up the creation of the entity tables because they require the
transformed dfs to pull ids and fields from. Close issue #159

also in the great migratoin, I inadvertantly removed the coalmine id
from the frc table. Instead of needing the loaed coalmine table, now it
uses a unique index for the id in both tables. Close issue #157