Make a boiler generator assn table
(!) Instead of having a more complete bga be in the post etl process,
now it is incorporated as a datatable.

This meant repurposing the bga infrastructure to take transformed_dfs as
an input instead of the pudl output object.

I extensively checked the difference between bga using the output object
and bga using the transform_dfs. Most of the issues ended up being data
type issues. I was able to individually fix those issues and came to a
comparable outcome. Bga using the dfs has ~1000 more records, most of
which are 2016 records that were being discarded in some way through the
output fucntions.

The other outcome needed was to contrain the years in which we are
creating bga. In the output functions we wanted to extend the bga,
assuming they didn't change too much so that we could get a more
extensive mcoe output, although we didn't want to put that date
extention into the datatables.