First attempt to specify a PUDL conda environment
This environment.yml file specifies versions for packages that are
required for PUDL to work, as well as a couple of outside of conda
packages (dbfread and sqlalchemy-postgres-copy) that get installed with
pip. The hope is that we can all know we're using similar versions of
the packages, and not unexpectedly encounter errors that aren't the
fault of the code -- but of different versions of the underlying tools
we're using.

To create an environment with this file, you do:

conda env create -f environment.yml

To activate that environment (whose name is pudl) you then do:

source activate pudl

Seems like the prefix shouldn't be in there, given that it's user
specific. See this page for more information:

https://conda.io/docs/user-guide/tasks/manage-environments.html