Modify EPA CEMS hourly ETL and testing
Made it possible to pull EPA CEMS data into the datastore by state, as
well as by year to ease testing and targeted usage. Did the same for the
database ETL.

Created a couple of basic ETL tests that are hopefully compatible with
Travis CI -- they pull down a minimal amount of data into a local
datastore (2016 data for all data sources, plus 2012 data for EIA 860
and FERC Form 1 -- as neither of those work with just 2016 data... need
to fix that).

Mostly the pull request was amazing!  Thank you @karldw for all the
great additions! Still some details to work out on the CEMS database
definition I think, and then integration with the other data, but well
on the way to being usable.

Also changed the default ETL to pull in only Colorado EPA CEMS data, to
avoid defaulting to a multi-hour process. If you want all the states and
all the years, you can use --states all overnight.

Partially addresses issue #147 and #181